requirements.txt
matplotlib
numpy
pandas
pillow
plotly
seaborn
streamlit